Move around some Stop copy helper functions
* Common functions extracted into stop-registry/utils/copyEntityUtilities.ts
* InfoSpot mapping into stop-registry/utils/mapInfoSpotToInput.ts
* Quay mapping into stop-registry/utils/mapQuayToInput.ts
* StopPlace mappin into /stop-registry/stop-areas/stop-area-details/mapStopPlaceToInput.ts
  - This causes a funny import path in mapCreateCopyFormStateToInputs.ts
    but that will get fixed later in ther actual "Copy Stop as Quay" commit.
