+++ ui/src/components/stop-registry/utils/index.ts
@@ -1,1 +1,4 @@
+export * from './copyEntityUtilities';
+export * from './mapInfoSpotToInput';
+export * from './mapQuayToInput';
 export * from './useShowStopAreaOnMap';
