+++ ui/src/components/stop-registry/utils/mapInfoSpotToInput.ts
@@ -1,0 +1,20 @@
+import {
+  InfoSpotDetailsFragment,
+  StopRegistryInfoSpotInput,
+} from '../../../generated/graphql';
+import {
+  mapCompactOrNull,
+  omitIdAndTypeName,
+  omitTypeName,
+} from './copyEntityUtilities';
+
+export function mapInfoSpotToInput(
+  infoSpot: InfoSpotDetailsFragment,
+): StopRegistryInfoSpotInput {
+  return {
+    ...omitIdAndTypeName(infoSpot),
+    infoSpotLocations: null,
+    description: omitTypeName(infoSpot.description),
+    poster: mapCompactOrNull(infoSpot.poster, omitTypeName),
+  };
+}
